+++ system/modules/core/controllers/FrontendIndex.php
@@ -134,9 +134,15 @@ class FrontendIndex extends \Frontend
 			}
 
 			// Try to find a page matching the language parameter
-			elseif (($lang = \Input::get('language')) != '' && isset($arrLangs[$lang]))
+			elseif (($lang = \Input::get('language')) && isset($arrLangs[$lang]))
 			{
 				$objNewPage = $arrLangs[$lang];
+			}
+
+			// Use the fallback language (see #8142)
+			elseif (isset($arrLangs['*']))
+			{
+				$objNewPage = $arrLangs['*'];
 			}
 
 			// Store the page object
@@ -147,7 +153,7 @@ class FrontendIndex extends \Frontend
 		}
 
 		// Throw a 404 error if the page could not be found or the result is still ambiguous
-		if ($objPage === null || ($objPage instanceof \Model\Collection && $objPage->count() != 1))
+		if ($objPage === null || ($objPage instanceof \Model\Collection && $objPage->count() > 1))
 		{
 			$this->User->authenticate();
 			$objHandler = new $GLOBALS['TL_PTY']['error_404']();
@@ -161,11 +167,23 @@ class FrontendIndex extends \Frontend
 		}
 
 		// If the page has an alias, it can no longer be called via ID (see #7661)
-		if ($objPage->alias != '' && preg_match('#^' . $objPage->id . '[$/.]#', \Environment::get('relativeRequest')))
+		if ($objPage->alias != '')
 		{
-			$this->User->authenticate();
-			$objHandler = new $GLOBALS['TL_PTY']['error_404']();
-			$objHandler->generate($pageId);
+			if (\Config::get('addLanguageToUrl'))
+			{
+				$regex = '#^[a-z]{2}(-[A-Z]{2})?/' . $objPage->id . '[$/.]#';
+			}
+			else
+			{
+				$regex = '#^' . $objPage->id . '[$/.]#';
+			}
+
+			if (preg_match($regex, \Environment::get('request')))
+			{
+				$this->User->authenticate();
+				$objHandler = new $GLOBALS['TL_PTY']['error_404']();
+				$objHandler->generate($pageId);
+			}
 		}
 
 		// Load a website root page object (will redirect to the first active regular page)