+++ system/modules/core/modules/ModuleCustomnav.php
@@ -166,7 +166,7 @@ class ModuleCustomnav extends \Module
 					$row['pageTitle'] = specialchars($objModel->pageTitle, true);
 					$row['link'] = $objModel->title;
 					$row['href'] = $href;
-					$row['nofollow'] = (strncmp($objModel->robots, 'noindex', 7) === 0);
+					$row['nofollow'] = (strncmp($objModel->robots, 'noindex,nofollow', 16) === 0);
 					$row['target'] = '';
 					$row['description'] = str_replace(array("\n", "\r"), array(' ' , ''), $objModel->description);
 
@@ -192,7 +192,7 @@ class ModuleCustomnav extends \Module
 					$row['pageTitle'] = specialchars($objModel->pageTitle, true);
 					$row['link'] = $objModel->title;
 					$row['href'] = $href;
-					$row['nofollow'] = (strncmp($objModel->robots, 'noindex', 7) === 0);
+					$row['nofollow'] = (strncmp($objModel->robots, 'noindex,nofollow', 16) === 0);
 					$row['target'] = '';
 					$row['description'] = str_replace(array("\n", "\r"), array(' ' , ''), $objModel->description);